+++ Changelog.md
@@ -4,6 +4,7 @@
 - New: Netherrack Mine and End Stone Mine
 - New: The damage dealt by a Laser Block containing a Harming Module can now be changed with a config setting
 - New: More modded wooden chests can now be converted to a Password-protected Chest
+- New: Crystal Quartz Bricks, Smooth Crystal Quartz, Smooth Crystal Quartz Stairs, Smooth Crystal Quartz Slab, and their reinforced variants
 - API: Changed IPasswordConvertible#getOriginalBlock to IPasswordConvertible#isValidStateForConversion for finer control over what can be converted
 - API: New methods IModuleInventory#isAllowed and IModuleInventory#isDenied to check whether an entity is listed on an allowlist or denylist module respectively
 - API: Moved and renamed Owner#isOwner methods to IPasswordProtected#isOwnedBy
@@ -16,7 +17,7 @@
 - Fix: Using /kill to remove a camera entity, or removing it by other non-standard means, does not update the camera block properly
 - Fix: Error when unbinding the last mine/sentry from a Mine/Sentry Remote Access Tool
 - Fix: Placing an Inventory Scanner facing a disguised scanner does not update the placed scanner's disguise
-- Fix: The Cage Trap and Password-protected Furnace/Smoker/Blast Furnace are black instead of see through
+- Fix: The Cage Trap and Password-protected Furnace/Smoker/Blast Furnace are black instead of see-through
 - Misc.: The minimum required Forge version is now 44.0.11
 
 --------------------------Changelog for v1.9.4-beta1 of SecurityCraft--------------------------